Use trade's direct rate when selling cc for base_currency. Fixes #12
Previously, when selling any currency (other than base_currency), the historic rate provided
by a CurrencyRelation object would be used to calculate profits. In case of a trade directly
between a cryptocurrency and base_currency, a more specific rate is known (the direct rate of
this trade) than relying on the averaged historic data. This more specific rate will be used now.